Add 'matching benchmark.py' script
This script uses the 'Rule.generate()' method to generate strings
for a rule and can be used benchmark how long it takes to match a
rule N times.

There is a default rule and other rules can be specified with the
'-r' argument. Profiling with 'cProfile' can be enabled with -p.